refactor (NuGettier.Upm): replace xxHash computation library Standart.Hash.xxHash with System.IO.Hashing (refactor/use-system.io.hashing-xxhash) [#171]
- build (NuGettier.Upm): add dependency on System.IO.Hashing v8.0.0
- build (NuGettier.Upm): remove dependency on Standart.Hash.xxHash
- refactor (NuGettier.Upm): replace xxHash computation library Standart.Hash.xxHash with System.IO.Hashing

rationale:
Guids, i.e. hashes generated with Standart.Hash.xxHash were getting rejected by Unity as
```
xxx.meta does not have a valid GUID and its corresponding Asset file will be ignored.
If this file is not malformed, please add a GUID, or delete the .meta file and it will be recreated correctly.

#0 GetStacktrace(int)
#1 DebugStringToFile(DebugStringToFileData const&)
#2 BuildReporting::LogRecordedErrorsToConsole(BuildReporting::BuildReport const&)
#3 Application::ParseARGVCommands()
#4 Application::FinishLoadingProject()
#5 NoGraphicsMain()
#6 EditorMain(int, char const**)
#7 main
#8 start
```

- build (NuGettier.Upm): add dependency on System.IO.Hashing v8.0.0
- build (NuGettier.Upm): remove dependency on Standart.Hash.xxHash
- refactor (NuGettier.Upm): replace xxHash computation library Standart.Hash.xxHash with System.IO.Hashing

rationale:
Guids, i.e. hashes generated with Standart.Hash.xxHash were getting rejected by Unity as
```
xxx.meta does not have a valid GUID and its corresponding Asset file will be ignored.
If this file is not malformed, please add a GUID, or delete the .meta file and it will be recreated correctly.

#0 GetStacktrace(int)
#1 DebugStringToFile(DebugStringToFileData const&)
#2 BuildReporting::LogRecordedErrorsToConsole(BuildReporting::BuildReport const&)
#3 Application::ParseARGVCommands()
#4 Application::FinishLoadingProject()
#5 NoGraphicsMain()
#6 EditorMain(int, char const**)
#7 main
#8 start
```